backend/pyproject.toml
[project]
name = "synsc-delphi"
version = "2.0.0"
description = "Open-source MCP server for indexing GitHub repositories, research papers, and HuggingFace datasets — runs fully local."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "InkVell" }
]
keywords = ["mcp", "github", "research", "papers", "code-search", "ai", "semantic-search", "cursor", "arxiv", "huggingface", "datasets"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # ===========================================
    # MCP Framework
    # ===========================================
    "mcp[cli]>=1.4.1",

    # ===========================================
    # Vector Search & Embeddings (all local)
    # ===========================================
    "faiss-cpu>=1.9.0",
    "google-genai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.40,<4.50",
    "torch>=2.0.0",

    # ===========================================
    # Database (local PostgreSQL + pgvector)
    # ===========================================
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",
    "alembic>=1.13.0",

    # ===========================================
    # Auth & Security
    # ===========================================
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",

    # ===========================================
    # Validation & Data
    # ===========================================
    "pydantic>=2.7.2",
    "numpy>=1.24.0",

    # ===========================================
    # Token counting
    # ===========================================
    "tiktoken>=0.7.0",

    # ===========================================
    # Logging
    # ===========================================
    "structlog>=24.1.0",

    # ===========================================
    # Rate Limiting
    # ===========================================
    "slowapi>=0.1.9",

    # ===========================================
    # HTTP client & server
    # ===========================================
    "httpx>=0.28.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # ===========================================
    # Git operations
    # ===========================================
    "dulwich>=0.22.0",

    # ===========================================
    # AST parsing for symbol extraction
    # ===========================================
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.22.0",
    "tree-sitter-javascript>=0.22.0",
    "tree-sitter-typescript>=0.22.0",

    # ===========================================
    # PDF & Research Paper processing
    # ===========================================
    "pymupdf>=1.23.0",
    "arxiv>=2.1.0",
    "markdownify>=1.1.0",
    "scikit-learn>=1.3.0",

    # ===========================================
    # HuggingFace Hub
    # ===========================================
    "huggingface-hub>=0.20.0",

    # ===========================================
    # Configuration
    # ===========================================
    "pyyaml>=6.0.1",

    # ===========================================
    # Misc
    # ===========================================
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "pytest-socket>=0.7.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
# Main CLI entry points
synsc-context = "synsc.cli:main"
synsc-context-mcp = "synsc.api.mcp_server:run_server"
synsc-context-http = "synsc.api.http_server:run_http_server"

# Lightweight stdio proxy for uvx usage
synsc-context-proxy = "synsc.mcp_stdio_proxy:main"

# Worker
synsc-context-worker = "synsc.workers.indexing_worker:run_worker"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["synsc"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["synsc"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q -v --durations=5"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = []
